- Refactored addAttributes and map functions into LdapAttributeUtils class. - Added configuration options for attribute mappings to LdapServerProperties.
